refactor: refactor and generalize single and staking validators
BREAKING CHANGE:
