+++ app/Main.hs
@@ -1,50 +1,25 @@
-{-# OPTIONS_GHC -Wno-unused-imports #-}
-
 module Main (main) where
 
 import Data.Aeson (KeyValue ((.=)), object)
 import Data.Aeson.Encode.Pretty (encodePretty)
-import Data.Bifunctor (
-  first,
- )
+import Data.Bifunctor (first)
 import Data.ByteString.Base16 qualified as Base16
 import Data.ByteString.Lazy qualified as LBS
-import Data.Default (
-  def,
- )
-import Data.Text (
-  Text,
-  pack,
- )
+import Data.Text (Text, pack)
 import Data.Text.Encoding qualified as Text
 import System.Console.ANSI (Color (..), ColorIntensity (..), ConsoleLayer (..), SGR (..), setSGR)
 
 import Cardano.Binary qualified as CBOR
-import PlutusLedgerApi.V1.Address (scriptHashAddress)
-import PlutusLedgerApi.V2 (
-  Data,
-  ExBudget,
- )
+import PlutusLedgerApi.V2 (Data, ExBudget)
 
-import Plutarch (
-  Config (Config),
-  TracingMode (DoTracing),
- )
-import Plutarch.Api.V1.Value (padaSymbol)
-import Plutarch.Evaluate (
-  evalScript,
- )
+import Plutarch.Evaluate (evalScript)
 import Plutarch.Prelude
 import Plutarch.Script (Script, serialiseScript)
-import Ply.Plutarch (
-  writeTypedScript,
- )
-import "liqwid-plutarch-extra" Plutarch.Extra.Script (
-  applyArguments,
- )
+import "liqwid-plutarch-extra" Plutarch.Extra.Script (applyArguments)
 
+import BatchValidator (smartHandleRouteValidatorW)
 import Compilation
-import SmartHandles
+import MinSwap.AdaToMin
 
 encodeSerialiseCBOR :: Script -> Text
 encodeSerialiseCBOR = Text.decodeUtf8 . Base16.encode . CBOR.serialize' . serialiseScript
@@ -64,10 +39,9 @@ writePlutusScript title filepath term = do
   case evalT term of
     Left e -> putStrLn (show e)
     Right (script, _, _) -> do
-      let
-        scriptType = "PlutusScriptV2" :: String
-        plutusJson = object ["type" .= scriptType, "description" .= title, "cborHex" .= encodeSerialiseCBOR script]
-        content = encodePretty plutusJson
+      let scriptType = "PlutusScriptV2" :: String
+          plutusJson = object ["type" .= scriptType, "description" .= title, "cborHex" .= encodeSerialiseCBOR script]
+          content = encodePretty plutusJson
       LBS.writeFile filepath content
 
 main :: IO ()
@@ -76,13 +50,13 @@ main = do
   putStrLn "Exporting Plutarch scripts..."
   setSGR [Reset]
 
-  writePlutusScript "Smart Handle" "./compiled/smartHandleSimple.json" psmartHandleValidatorW
+  writePlutusScript "Smart Handle" "./compiled/smartHandleSimple.json" psingleValidator
   putStrLn "Exported smart handle validator"
 
   writePlutusScript "Smart Handle Router" "./compiled/smartHandleRouter.json" smartHandleRouteValidatorW
   putStrLn "Exported smart handle router validator"
 
-  writePlutusScript "Smart Handle Router" "./compiled/smartHandleStake.json" smartHandleStakeValidatorW
+  writePlutusScript "Smart Handle Router" "./compiled/smartHandleStake.json" pstakeValidator
   putStrLn "Exported smart handle stake validator"
 
   setSGR [SetColor Foreground Vivid Green]
